Extracted tab to its own component + refactor
